pyproject.toml
[tool.poetry]
name = "ide_core"
version = "0.1.0"
description = ""
authors = ["Jatin Katyal <jatin.katyal13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
requests = "^2.25.1"
django-rest-framework = "^0.1.0"
python-dotenv = "^0.17.1"
django-cors-headers = "^3.7.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "19.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
